Refactor `CallToAction` inner block
